# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateSLARequest": {
        "properties": {
          "alert_channels": {
            "items": {
              "$ref": "#/components/schemas/AlertChannel"
            },
            "title": "Alert Channels",
            "type": "array"
          },
          "check_frequency": {
            "$ref": "#/components/schemas/SLACheckFrequency",
            "default": "hourly"
          },
          "max_carbon_intensity_gco2_kwh": {
            "default": 100.0,
            "minimum": 0.0,
            "title": "Max Carbon Intensity Gco2 Kwh",
            "type": "number"
          },
          "min_renewable_percentage": {
            "default": 50.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Min Renewable Percentage",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "providers": {
            "items": {
              "type": "string"
            },
            "title": "Providers",
            "type": "array"
          },
          "regions": {
            "items": {
              "type": "string"
            },
            "title": "Regions",
            "type": "array"
          },
          "webhook_url": {
            "default": "",
            "title": "Webhook Url",
            "type": "string"
          }
        },
        "required": [
          "org_id",
          "name"
        ],
        "title": "CreateSLARequest",
        "type": "object"
      },
...
      "GreenSLA": {
        "description": "A carbon SLA definition \u2014 the target a customer commits to.",
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "alert_channels": {
            "items": {
              "$ref": "#/components/schemas/AlertChannel"
            },
            "title": "Alert Channels",
            "type": "array"
          },
          "check_frequency": {
            "$ref": "#/components/schemas/SLACheckFrequency",
            "default": "hourly"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "max_carbon_intensity_gco2_kwh": {
            "description": "Maximum allowed carbon intensity (gCO2/kWh)",
            "minimum": 0.0,
            "title": "Max Carbon Intensity Gco2 Kwh",
            "type": "number"
          },
          "min_renewable_percentage": {
            "description": "Minimum renewable energy percentage",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Min Renewable Percentage",
            "type": "number"
          },
          "name": {
            "description": "Human-readable SLA name",
            "title": "Name",
            "type": "string"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "providers": {
            "description": "Cloud providers covered by this SLA",
            "items": {
              "type": "string"
            },
            "title": "Providers",
            "type": "array"
          },
          "regions": {
            "description": "Specific regions to monitor (empty = all regions for the providers)",
            "items": {
              "type": "string"
            },
            "title": "Regions",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "webhook_url": {
            "default": "",
            "title": "Webhook Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "org_id",
          "name",
          "max_carbon_intensity_gco2_kwh",
          "min_renewable_percentage",
          "created_at",
          "updated_at"
        ],
        "title": "GreenSLA",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "A free, open API for the real-time carbon intensity of cloud regions. No account or API key required; usage is rate-limited so it stays responsive for everyone.\n\n## Carbon Data API\n- **8 live grid-operator integrations** (UK, EIA, OpenElectricity/AEMO, IESO/AESO, Taipower, GridStatus, ENTSO-E, Electricity Maps) plus labeled heuristic and weather-based estimates, cascading\n- **Grid carbon intensity** for 75+ cloud regions; every response is tagged with its `source`\n- **Batch queries** for multiple regions in a single call\n\n## Compliance Reporting\n- **GHG-Protocol-structured** Scope 2 (location-based) + Scope 3 Cat 1 reporting\n- Drafts aimed at **CSRD / SEC Climate / SB 253** workflows (not an assured report)\n- **EU Taxonomy** eligibility flag (simplified screening)\n\n## Green SLA Monitoring (Beta)\n- **Define carbon targets** \u2014 max gCO2/kWh, min renewable %\n- **On-demand + background checks** (in-memory state)\n- **Attestation-style summary reports**\n\n## Carbon-Aware Scheduling\n- **Find optimal time windows** for batch jobs, CI/CD, ML training\n- **Multi-region evaluation** across AWS, GCP, Azure\n- **Three strategies** \u2014 lowest carbon, highest renewable, balanced\n- Returns a recommendation \u2014 it doesn't run or defer your workload itself",
    "title": "Carbon Lens",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sla/create": {
      "post": {
        "description": "Create a new Green SLA definition.",
        "operationId": "create_sla_api_v1_sla_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSLARequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenSLA"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Sla",
        "tags": [
          "SLA Monitoring"
        ]
      }
    },
    "/api/v1/sla/list": {
      "get": {
        "description": "List all SLAs for an organization.",
        "operationId": "list_slas_api_v1_sla_list_get",
        "parameters": [
          {
            "in": "query",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SLASummary"
                  },
                  "title": "Response List Slas Api V1 Sla List Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Slas",
        "tags": [
          "SLA Monitoring"
        ]
      }
    },
    "/api/v1/sla/monitor/alerts": {
      "get": {
        "description": "List recent SLA breach alerts.",
        "operationId": "list_alerts_api_v1_sla_monitor_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertEvent"
                  },
                  "title": "Response List Alerts Api V1 Sla Monitor Alerts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Alerts",
        "tags": [
          "SLA Monitoring"
        ]
      }
    },
    "/api/v1/sla/monitor/run": {
      "post": {
        "description": "Run every SLA check that's due now and persist the results.\n\nThis is the durable, scale-to-zero-friendly path: a scheduler (the GitHub\nActions cron in .github/workflows/sla-monitor.yml) POSTs here on a cadence, so\nchecks happen and persist even when the API would otherwise be idle -- unlike\nthe in-process monitor loop, which only runs while the service is awake.\nAdmin-only (X-API-Key must match CARBON_LENS_ADMIN_SECRET).",
        "operationId": "run_due_checks_api_v1_sla_monitor_run_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Due Checks Api V1 Sla Monitor Run Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Run Due Checks",
        "tags": [
          "SLA Monitoring"
        ]
      }
    },
    "/api/v1/sla/monitor/start": {
      "post": {
        "description": "Start the background SLA monitor for an organization's SLAs.\n\nNote: the in-process monitor only runs while the API is awake. For durable,\nscheduled checking on a scale-to-zero host, use POST /monitor/run from a cron.",
        "operationId": "start_monitor_api_v1_sla_monitor_start_post",
        "parameters": [
          {
            "in": "query",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Start Monitor Api V1 Sla Monitor Start Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Start Monitor",
        "tags": [
          "SLA Monitoring"
        ]
      }
    },
    "/api/v1/sla/monitor/status": {
      "get": {
        "description": "Get the current status of the SLA monitor.",
        "operationId": "monitor_status_api_v1_sla_monitor_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Monitor Status Api V1 Sla Monitor Status Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Monitor Status",
        "tags": [
          "SLA Monitoring"
        ]
      }
    },
    "/api/v1/sla/monitor/stop": {
      "post": {
        "description": "Stop the background SLA monitor.",
        "operationId": "stop_monitor_api_v1_sla_monitor_stop_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Stop Monitor Api V1 Sla Monitor Stop Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Stop Monitor",
        "tags": [
          "SLA Monitoring"
        ]
      }
    },
    "/api/v1/sla/{sla_id}": {
      "delete": {
        "description": "Delete an SLA.",
        "operationId": "delete_sla_api_v1_sla__sla_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "sla_id",
            "required": true,
            "schema": {
              "title": "Sla Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Sla Api V1 Sla  Sla Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Delete Sla",
        "tags": [
          "SLA Monitoring"
        ]
      },
      "get": {
        "description": "Get an SLA definition by ID.",
        "operationId": "get_sla_api_v1_sla__sla_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "sla_id",
            "required": true,
            "schema": {
              "title": "Sla Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenSLA"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Sla",
        "tags": [
          "SLA Monitoring"
        ]
      },
      "put": {
        "description": "Update an existing SLA.",
        "operationId": "update_sla_api_v1_sla__sla_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "sla_id",
            "required": true,
            "schema": {
              "title": "Sla Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSLARequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenSLA"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Update Sla",
        "tags": [
          "SLA Monitoring"
        ]
      }
    },
    "/api/v1/sla/{sla_id}/check": {
      "post": {
        "description": "Run an on-demand SLA compliance check.\n\nFetches live carbon data for all monitored regions and evaluates\nagainst the SLA thresholds.",
        "operationId": "check_sla_api_v1_sla__sla_id__check_post",
        "parameters": [
          {
            "in": "path",
            "name": "sla_id",
            "required": true,
            "schema": {
              "title": "Sla Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SLACheck"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Check Sla",
        "tags": [
          "SLA Monitoring"
        ]
      }
    },
    "/api/v1/sla/{sla_id}/checks": {
      "get": {
        "description": "List recent compliance checks for an SLA.",
        "operationId": "list_checks_api_v1_sla__sla_id__checks_get",
        "parameters": [
          {
            "in": "path",
            "name": "sla_id",
            "required": true,
            "schema": {
              "title": "Sla Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SLACheck"
                  },
                  "title": "Response List Checks Api V1 Sla  Sla Id  Checks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Checks",
        "tags": [
          "SLA Monitoring"
        ]
      }
    },
    "/api/v1/sla/{sla_id}/report": {
      "post": {
        "description": "Generate an attestation report for an SLA over a period.\n\nUses stored compliance checks to build the report. If no checks exist\nfor the period, runs a fresh check first.",
        "operationId": "generate_report_api_v1_sla__sla_id__report_post",
        "parameters": [
          {
            "in": "path",
            "name": "sla_id",
            "required": true,
            "schema": {
              "title": "Sla Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/carbonlens__sla__routes__GenerateReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SLAReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Generate Report",
        "tags": [
          "SLA Monitoring"
        ]
      }
    },
    "/api/v1/sla/{sla_id}/reports": {
      "get": {
        "description": "List all attestation reports for an SLA.",
        "operationId": "list_reports_api_v1_sla__sla_id__reports_get",
        "parameters": [
          {
            "in": "path",
            "name": "sla_id",
            "required": true,
            "schema": {
              "title": "Sla Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SLAReport"
                  },
                  "title": "Response List Reports Api V1 Sla  Sla Id  Reports Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Reports",
        "tags": [
          "SLA Monitoring"
        ]
      }
    },
    "/api/v1/sla/{sla_id}/status": {
      "get": {
        "description": "Get the most recent compliance check for an SLA.",
        "operationId": "get_sla_status_api_v1_sla__sla_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "sla_id",
            "required": true,
            "schema": {
              "title": "Sla Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SLACheck"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Sla Status Api V1 Sla  Sla Id  Status Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Sla Status",
        "tags": [
          "SLA Monitoring"
        ]
      }
    },
...
  },
...
}